Ajout de l’importation du module fs pour gérer la suppression des fichiers. Mise à jour de la fonction deleteBook : vérification que l’utilisateur authentifié est bien le propriétaire du Book, extraction du nom de fichier depuis imageUrl via split('/images/'), suppression du fichier image avec fs.unlink pour éviter les fichiers orphelins et suppression du Book dans la base de données une fois le fichier supprimé.

+++ controllers/bookCtrl.js
@@ -1,4 +1,7 @@
 const Book = require('../models/book');
+const fs = require('fs');
+//l nous donne accès aux fonctions qui nous permettent de modifier le système de fichiers
+//, y compris aux fonctions permettant de supprimer les fichiers.
 
 exports.createBook = (req, res, next) => {
     const bookObject = JSON.parse(req.body.book);
@@ -46,16 +49,61 @@ exports.ratingBook = (req, res, next) => {
     .catch(error => res.status(400).json({ error }));
 };
 
+exports.bestRatingBook = (req, res, next) => {
+  Book.find()
+    .then(books => {
+        const sortBooks = books.sort((a, b) => b.averageRating - a.averageRating);
+        const bestThree = sortBooks.slice(0, 3);
+        res.status(200).json(bestThree);
+    })
+    .catch(error => res.status(400).json({ error }));
+};
+
+exports.getOneBook = (req, res, next) => {
+   Book.findOne({_id: req.params.id})
+    .then(book => res.status(200).json(book))
+    .catch(error => res.status(404).json({ error }));
+};
+
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({_id: req.params.id}, { ...req.body, _id:req.params.id })
-  .then(() => res.status(201).json({ message: 'Book updated successfully!'}))
-  .catch(error => res.status(400).json({ error }));
+  const bookObject = req.file ? {
+       ...JSON.parse(req.body.book),
+       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+   } : { ...req.body};
+ 
+   delete bookObject.userId;
+   Book.findOne({_id: req.params.id})
+       .then((book) => {
+           if (book.userId != req.auth.userId) {
+               res.status(401).json({ message : 'Not authorized'});
+           } else {
+               Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+               .then(() => res.status(200).json({message : 'livre modifié!'}))
+               .catch(error => res.status(401).json({ error }));
+           }
+       })
+       .catch((error) => {
+           res.status(400).json({ error });
+       });
 };
 
 exports.deleteBook = (req, res, next) => {  
-    Book.deleteOne({_id: req.params.id})
-    .then(() => res.status(200).json({ message: 'Deleted!'}))
-    .catch(error => res.status(400).json({ error }));
+    Book.findOne({ _id: req.params.id})
+       .then(book => {
+           if (book.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = book.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Book.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
 
 exports.getAllBook = (req, res, next) => {
